% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EP_Draft_Scraper.R
\name{EP_Draft_Scraper}
\alias{EP_Draft_Scraper}
\title{Draft URL Scraper}
\usage{
EP_Draft_Scraper(Data, Agerange = c(17, 25), draft.year = T,
  draft.pick = T, round = T, Agerel = "9/15", Goalie = F,
  position = T, shoots = T, Stats = c("S", "Team", "League", "GP",
  "G", "A", "TP", "PIM", "+/-", "sv\%", "GAA"), place.birth = T,
  pbsep = T, Country = T, height = T, weight = T, date.birth = T,
  dbsep = T, drafted.team = T, reg.playoffs = "R")
}
\arguments{
\item{Data}{the url of the draft year webpage on eliteprospects.com.}

\item{Agerange}{2-length vector for the age ranges desired. first number is lower-bound, second number is upper-bound. This will be applied to all players.}

\item{draft.year}{Boolean about whether to include the draft year of the player.}

\item{draft.pick}{Boolean about whether to include the draft pick used on the player.}

\item{round}{Boolean about whether to include the round the player was drafted in.}

\item{Agerel}{This determines how the exact age included in the data frame is calculated. Default is 9/15 due to this being the cutoff data for draft eligibility. For example, during the 2019-2020 season, if Agerel is 9/15, the age in the data frame will represent the players age on 9/15/2020. The offset here is so that for a player's first draft eligible year, their age will be >18.}

\item{Goalie}{Boolean about whether information on goalies is wanted. If true, output will be a list of data frames, with one for players, and one for goalies. Currently this doesn't do anything, as there isn't a goalie scraper function built out yet.}

\item{position}{Boolean about whether to include the position of the player.}

\item{shoots}{Boolean about whether to include the handedness of the player.}

\item{Stats}{vector of the wanted stats. 
S - Season
Team - Team
League - League
GP - Games Played
G - Goals
A - Assists
TP - Total Points
PIM - Penalties in Minutes
+/- - Plus/Minus
sv% - Save Percentage (for goalies)
GAA - Goals Against Average (for goalies)}

\item{place.birth}{Boolean about whether to include the birthplace of the player.}

\item{pbsep}{Boolean about whether the birthplace should be split into Country, State, and City. place.birth has to be true for this to matter.}

\item{Country}{Boolean about whether to include the country the player represents (or would represent) in international tournaments. Currently just grabs the first if a player has multiple, but eliteprospects is generally good about putting the correct one first.}

\item{height}{Boolean about whether to include the height of the player. This is in centimeters.}

\item{weight}{Boolean about whether to include the weight of the player. This is in pounds.}

\item{date.birth}{Boolean about whether to include the date of birth of the player.}

\item{dbsep}{Boolean about whether the date of birth should be seperated into year, month, and day. date.birth has to be true for this to matter.}

\item{drafted.team}{Boolean about whether to include the team who drafted the player.}

\item{reg.playoffs}{Determines if regular season data, playoff data, or both will be returned for all the players. Currently only 'R' works.}
}
\value{
data frame (or list of data frames) with data from all the players drafted in the given year.
}
\description{
Returns data from all players in the given draft year url
}
