% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EP_League_Links.R
\name{EP_League_Links}
\alias{EP_League_Links}
\title{League URL gatherer}
\usage{
EP_League_Links(Data, undrafted = T)
}
\arguments{
\item{Data}{the url of the league and year stats page from eliteprospects.}

\item{undrafted}{Boolean about whether to return only undrafted players, or any players (regardless of whether or not they were drafted).}
}
\value{
A vector of player eliteprospects URLs
}
\description{
Reutrns a list of URLs of the players from a league stats URL.
}
