% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EP_Name_Search.R
\name{EP_Name_Search}
\alias{EP_Name_Search}
\title{Name Search Scrapper}
\usage{
EP_Name_Search(Name)
}
\arguments{
\item{Name}{A string of the desired player's name}
}
\value{
A vector of player eliteprospects URLs with the same name
}
\description{
Returns a list of URLs of the players from a league stats URL.
}
